infrastructure/MetaTrader5/assets/requirements.txt
numpy==1.21.4 #==1.19.5 # ==1.18.2
MetaTrader5 # ==5.0.31
rpyc # ==5.2.3
pyparsing==2.4.7
# mt5linux
pandas==1.4.1
# 
vncdotool 
python-dotenv